+++ src/services/Gql.php
@@ -24,7 +24,7 @@ use craft\gql\ArgumentManager;
 use craft\gql\base\Directive;
 use craft\gql\base\GeneratorInterface;
 use craft\gql\base\InterfaceType;
-use craft\gql\base\NonGeneratedTypeInterface;
+use craft\gql\base\SingularTypeInterface;
 use craft\gql\directives\FormatDateTime;
 use craft\gql\directives\Markdown;
 use craft\gql\directives\ParseRefs;
@@ -1299,7 +1299,7 @@ class Gql extends Component
         $this->trigger(self::EVENT_REGISTER_GQL_TYPES, $event);
 
         foreach ($event->types as $type) {
-            /** @var NonGeneratedTypeInterface $type */
+            /** @var SingularTypeInterface $type */
             TypeLoader::registerType($type::getName(), $type . '::getType');
         }