+++ CHANGELOG.md
@@ -33,7 +33,7 @@
 - Added `craft\errors\MissingVolumeFolderException`.
 - Added `craft\fieldlayoutelements\BaseNativeField`, which replaces `craft\fieldlayoutelements\StandardField`.
 - Added `craft\fieldlayoutelements\TextField`, which replaces `craft\fieldlayoutelements\StandardTextField`.
-- Added `craft\gql\base\NonGeneratedTypeInterface`.
+- Added `craft\gql\base\SingularTypeInterface`.
 - Added `craft\helpers\Assets::downloadFile()`.
 - Added `craft\helpers\Cp::dateFieldHtml()`.
 - Added `craft\helpers\Cp::dateHtml()`.